Extract CollapsibleMarkdownCard shared by Soul and README cards
EntitySoulCard and EntityReadmeCard duplicated the InfoCard shell,
height-limited Collapse, fade overlay, toggle button, and the
ResizeObserver/MutationObserver wiring used to remeasure async markdown.
Move all of that into a single CollapsibleMarkdownCard under
components/UI, and reduce the two entity cards to thin wrappers that
pass their hook output in. README's NotFoundError message is preserved
via an optional renderError prop on the shared component.
