+++ plugins/gs/src/components/UI/CollapsibleMarkdownCard/index.ts
@@ -1,0 +1,1 @@
+export { CollapsibleMarkdownCard } from './CollapsibleMarkdownCard';
