+++ plugins/gs/src/components/UI/index.ts
@@ -4,6 +4,7 @@ export { Account } from './Account';
 export { ApplicationLink } from './ApplicationLink';
 export { ClusterLink } from './ClusterLink';
 export { CodeBlock } from './CodeBlock';
+export { CollapsibleMarkdownCard } from './CollapsibleMarkdownCard';
 export { ColorWrapper } from './ColorWrapper';
 export { ContentRow } from './ContentRow';
 export { DateComponent } from './Date';
