+++ plugins/gs/src/components/catalog/EntityReadmeCard/EntityReadmeCard.tsx
@@ -1,48 +1,12 @@
-import { useState, useRef, useEffect, useCallback } from 'react';
-import { MarkdownContent, Progress } from '@backstage/core-components';
-import { InfoCard } from '@giantswarm/backstage-plugin-ui-react';
-import {
-  Box,
-  Button,
-  Collapse,
-  makeStyles,
-  Typography,
-} from '@material-ui/core';
-import DescriptionIcon from '@material-ui/icons/Description';
+import { Typography } from '@material-ui/core';
 import { useCurrentEntityChart } from '../EntityChartContext';
 import { useHelmChartTags } from '../../hooks/useHelmChartTags';
 import { useHelmChartReadme } from '../../hooks/useHelmChartReadme';
 import { QueryClientProvider } from '../../QueryClientProvider';
 import { parseChartRef } from '../../utils/parseChartRef';
-
-const COLLAPSED_MAX_HEIGHT = 250;
-
-const useStyles = makeStyles(theme => ({
-  contentWrapper: {
-    position: 'relative',
-  },
-  fadeOverlay: {
-    position: 'absolute',
-    bottom: 0,
-    left: 0,
-    right: 0,
-    height: 80,
-    background: `linear-gradient(to bottom, transparent, ${theme.palette.background.paper})`,
-    pointerEvents: 'none',
-  },
-  toggleButton: {
-    display: 'flex',
-    justifyContent: 'center',
-    marginTop: theme.spacing(1),
-  },
-  titleIcon: {
-    marginRight: theme.spacing(1),
-    verticalAlign: 'middle',
-  },
-}));
+import { CollapsibleMarkdownCard } from '../../UI';
 
 const ReadmeCardContent = () => {
-  const classes = useStyles();
   const { selectedChart } = useCurrentEntityChart();
   const {
     latestStableVersion,
@@ -55,145 +19,29 @@ const ReadmeCardContent = () => {
     error: readmeError,
   } = useHelmChartReadme(selectedChart.ref, latestStableVersion ?? undefined);
 
-  const [expanded, setExpanded] = useState(false);
-  const [needsExpansion, setNeedsExpansion] = useState(false);
-  const [collapsedSize, setCollapsedSize] = useState(COLLAPSED_MAX_HEIGHT);
-  const contentRef = useRef<HTMLDivElement>(null);
-
-  const isLoading = isLoadingTags || isLoadingReadme;
-
-  const error = tagsError || readmeError;
-
-  const updateContentSize = useCallback(() => {
-    if (contentRef.current) {
-      const contentHeight = contentRef.current.scrollHeight;
-      setNeedsExpansion(contentHeight > COLLAPSED_MAX_HEIGHT);
-      setCollapsedSize(
-        contentHeight > COLLAPSED_MAX_HEIGHT
-          ? COLLAPSED_MAX_HEIGHT
-          : contentHeight,
-      );
-    }
-  }, []);
-
-  // Update size when readme content changes
-  useEffect(() => {
-    if (readme) {
-      updateContentSize();
-    }
-  }, [readme, updateContentSize]);
-
-  // Update size when viewport resizes
-  useEffect(() => {
-    if (!readme || !contentRef.current) {
-      return undefined;
-    }
-
-    const resizeObserver = new ResizeObserver(() => {
-      updateContentSize();
-    });
-
-    resizeObserver.observe(contentRef.current);
-
-    return () => {
-      resizeObserver.disconnect();
-    };
-  }, [readme, updateContentSize]);
-
-  // Detect when MarkdownContent finishes rendering its HTML inside the
-  // Collapse. The Collapse constrains the outer height, so ResizeObserver
-  // alone may not fire when content is injected asynchronously.
-  useEffect(() => {
-    if (!readme || !contentRef.current) {
-      return undefined;
-    }
-
-    const mutationObserver = new MutationObserver(() => {
-      updateContentSize();
-    });
-
-    mutationObserver.observe(contentRef.current, {
-      childList: true,
-      subtree: true,
-    });
-
-    return () => {
-      mutationObserver.disconnect();
-    };
-  }, [readme, updateContentSize]);
-
-  const handleToggle = () => {
-    setExpanded(prev => !prev);
-  };
-
-  const renderContent = () => {
-    if (isLoading) {
-      return <Progress />;
-    }
-
-    if (error) {
-      if (error.name === 'NotFoundError') {
-        const { repository } = parseChartRef(selectedChart.ref);
-        return (
-          <Typography variant="inherit" color="textSecondary">
-            The repository <code>{repository}</code> is not available in the
-            registry.
-          </Typography>
-        );
-      }
-      return <Typography color="error">{error.message}</Typography>;
-    }
-
-    if (!readme) {
+  const renderError = (error: Error) => {
+    if (error.name === 'NotFoundError') {
+      const { repository } = parseChartRef(selectedChart.ref);
       return (
         <Typography variant="inherit" color="textSecondary">
-          No README available.
+          The repository <code>{repository}</code> is not available in the
+          registry.
         </Typography>
       );
     }
-
-    return (
-      <>
-        <Box className={classes.contentWrapper}>
-          <Collapse
-            in={expanded || !needsExpansion}
-            collapsedSize={collapsedSize}
-          >
-            <div ref={contentRef}>
-              <MarkdownContent content={readme} />
-            </div>
-          </Collapse>
-          {needsExpansion && !expanded && (
-            <div className={classes.fadeOverlay} />
-          )}
-        </Box>
-        {needsExpansion && (
-          <Box className={classes.toggleButton}>
-            <Button
-              variant="text"
-              color="primary"
-              onClick={handleToggle}
-              size="small"
-            >
-              {expanded ? 'Show less' : 'Show full README'}
-            </Button>
-          </Box>
-        )}
-      </>
-    );
+    return undefined;
   };
 
   return (
-    <InfoCard
-      title={
-        <Box display="flex" alignItems="center">
-          <DescriptionIcon className={classes.titleIcon} />
-          README
-        </Box>
-      }
-    >
-      {renderContent()}
-    </InfoCard>
+    <CollapsibleMarkdownCard
+      title="README"
+      content={readme}
+      isLoading={isLoadingTags || isLoadingReadme}
+      error={tagsError || readmeError}
+      emptyMessage="No README available."
+      toggleLabels={{ expand: 'Show full README', collapse: 'Show less' }}
+      renderError={renderError}
+    />
   );
 };
